Refactor: estructura de carpetas de componentes
